Create request service
Done for convenience's sake. The service is injected as a context at the
root level so any subsequent component can use it as they deem fit. It feels
much better than manually importing the service everywhere.

It's also a singleton as different instance's won't really do anything.

Added some caching of the urls to avoid doing the same concat operations over
and over.